requirements.txt
requests == 2.27.1
urllib3 == 1.26.9
nose==1.3.7
Click==8.1.0
GitPython==3.1.27
PyYAML==5.4.1
cryptography==2.4.2
Fabric3>=1.1,<2.0
paramiko==2.4.3
invoke==1.7.1
